backend/requirements.txt
fastapi
uvicorn
sqlmodel
playwright
apscheduler
python-multipart
emails
jinja2
python-dotenv
